just-biomarkers/pyproject.toml
[project]
name = "just-biomarkers"
version = "0.1.0"
description = "Polars-based methylation clock computation engine"
readme = "README.md"
authors = [
    { name = "antonkulaga", email = "antonkulaga@gmail.com" }
]
requires-python = ">=3.11,<3.14"
dependencies = [
    "polars>=1.39.0",
    "typer>=0.24.1",
    "rich>=14.3.3",
    "pydantic>=2.0",
    "platformdirs>=4.0.0",
    "numpy>=2.4.3",
    "pylluminator-modern>=1.3; python_version >= '3.12'",
    "huggingface-hub>=1.8.0",
]

[project.scripts]
just-biomarkers = "just_biomarkers.cli:app"
biomarkers = "just_biomarkers.cli:app"

[build-system]
requires = ["uv_build>=0.10.0,<0.11.0"]
build-backend = "uv_build"
